refactor: move Errors to Result.Immutable namespace
