+++ src/Results.Immutable.Tests/Usings.cs
@@ -1,5 +1,3 @@
 global using Xunit;
-global using Results.Immutable;
-global using Results.Immutable.Metadata;
 global using FluentAssertions;
 global using System.Collections.Immutable;
